Make component InlineCode standalone
